updated addon loading
